Move `ValidatedFormField` to `fields.py`
