+++ cla_public/apps/callmeback/fields.py
@@ -195,3 +195,26 @@ class AvailabilityCheckerField(FormField):
 
     def scheduled_time(self):
         return self.form.scheduled_time().replace(tzinfo=pytz.utc)
+
+
+class ValidatedFormField(FormField):
+    def __init__(self, form_class, *args, **kwargs):
+        self._errors = []
+        self.validators = kwargs.pop('validators', [])
+
+        super(ValidatedFormField, self).__init__(
+            form_class, *args, **kwargs)
+
+    def validate(self, form, extra_validators=None):
+        if self._run_validation_chain(form, self.validators):
+            return len(self.errors) == 0
+        form_valid = self.form.validate()
+        return form_valid and len(self.errors) == 0
+
+    @property
+    def errors(self):
+        return self._errors + self.form.errors.items()
+
+    @errors.setter
+    def errors(self, _errors):
+        self._errors = _errors
